Make backfille[er] common_version hold version_t
The purpose of which is to provide a convenient branch_id, version_t
value for the backfillee to use when traversing the WAL.  It occurs to
me that the backfillee could recompute the branch_id from the
state_timestamp_t itself.  But since we throw away that value on the
backfiller side, it is less code and less work, and safer, to do
things this way.