Make totally_exhausted() just read from key_range instead of state
This is not proven to be a logically identical transformation of code.
Hypothetically the key range is empty in new_active_ranges and then we
start with an empty key_range that is ACTIVE.  I think that cannot be
the case, and even if it were, it would save a spurious read.